+++ tmf/org.eclipse.tracecompass.tmf.ui/src/org/eclipse/tracecompass/tmf/ui/widgets/timegraph/model/ILinkEvent.java
@@ -27,5 +27,5 @@ public interface ILinkEvent extends ITimeEvent {
      */
     ITimeGraphEntry getDestinationEntry();
 
-    int getValue();
+    long getValue();
 }
